AX: AXObjectCache::remove(Node&) can be called while performing a cache update causing m_deferred* vectors to be modified as we iterate over them https://bugs.webkit.org/show_bug.cgi?id=244421 rdar://98729717
Reviewed by Chris Fleizach.

Handling certain attribute changes in AXObjectCache::handleAttributeChange
can cause AXObjectCache::remove(Node&) to be called as a side effect.
For example, this can happen as a result of a `role` attribute change.
When this happens, the Node is removed from m_deferredAttributeChanges while
we are iterating over this vector.

This patch prevents this by making AXObjectCache::remove(Node&) bail
early before modifying any m_deferred* vector if we detect that we're in
the middle of a cache update. If this is true, these vectors will be
cleared anyways upon completion of their processing.

* LayoutTests/accessibility/multiple-attribute-change-crash-expected.txt: Added.
* LayoutTests/accessibility/multiple-attribute-change-crash.html: Added.
* LayoutTests/platform/ios/TestExpectations: Enable new test.
* LayoutTests/platform/glib/accessibility/multiple-attribute-change-crash-expected.txt: Added
* LayoutTests/platform/ios/accessibility/multiple-attribute-change-crash-expected.txt: Added.
* LayoutTests/platform/win/TestExpectations: Disable new test.
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::remove):
(WebCore::AXObjectCache::handleAttributeChange):
Fix unchecked null dereference of the input Element* in an AXLOG.

Canonical link: https://commits.webkit.org/253871@main