Atomics.waitAsync should be invocable from the main thread https://bugs.webkit.org/show_bug.cgi?id=250518
Reviewed by Keith Miller.

According to the current proposed spec [1], async atomic waits are not
restricted to threads with `AgentCanSuspend()`--so, Atomics.waitAsync should not
throw a TypeError for this reason. I've added a (very simple) test to this
effect.

So, we should delay the check for `isAtomicsWaitAllowedOnCurrentThread` until after we know
the wait is known to be synchronous.

If there's an implementation reason we shouldn't do this, let me know, but I
couldn't find one.

[1] https://tc39.es/proposal-atomics-wait-async/#sec-dowait

* LayoutTests/js/Atomics-waitasync-invocable-on-main-thread-expected.txt: Added.
* LayoutTests/js/Atomics-waitasync-invocable-on-main-thread.html: Added.
* Source/JavaScriptCore/runtime/AtomicsObject.cpp:
(JSC::atomicsWaitImpl):

Canonical link: https://commits.webkit.org/258856@main