[GStreamer][MSE] Append pipeline is counting streams incorrectly with new init segments https://bugs.webkit.org/show_bug.cgi?id=254664
Reviewed by Alicia Boya Garcia.

When you append another init segment we match GStreamer pads against tracks but we are not counting properly as we are
skipping the pads that are already linked we can end up with less pads than tracks. Besides it is against the spec. We
need to match the number of tracks and track IDs if there are more than one for a kind.

* LayoutTests/media/media-source/media-source-seek-detach-crash.html: Fail the test earlier instead of waiting for a
timeout.
* LayoutTests/platform/glib/TestExpectations: Unmark
media/encrypted-media/encrypted-media-append-encrypted-unencrypted.html as failure. It works now.
* Source/WebCore/platform/graphics/gstreamer/mse/AppendPipeline.cpp:
(WebCore::AppendPipeline::didReceiveInitializationSegment):
(WebCore::AppendPipeline::recycleTrackForPad):

Canonical link: https://commits.webkit.org/264676@main