Ensure that a single composition happens for each acc canvas or WebGL frame.
Rendering a new WebGL or accelerated canvas frame can request two compositions,
one due to the layerFlush and another one because of the new buffer pushed to
the proxies. This can cause problems to apps relying on rAF. This commit disables
the composition request that comes from pushing a new buffer from WebGL or the
acc canvas, and ensures that the only composition request comes from the layerFlush.