[GStreamer] Setting playbackRate=0 does not pause video for progressive video playback https://bugs.webkit.org/show_bug.cgi?id=267601
Reviewed by Philippe Normand.

Setting playbackRate=0 on a regular video doesn't pause the playback. Also, setting
playbackRate=1 (or any other value) on a paused video doesn't resume the playback.

See: https://github.com/WebPlatformForEmbedded/WPEWebKit/pull/1272

There's an m_playbackRatePausedState attribute in the curent code, but it only keeps track
of the desired state to be reported to the upper layers and also about if the video should
be paused/unpaused because of buffering. It's not working in practice to make the
playbackRate changes done from JavaScript effective in the pipeline.

This patch adds code to update the pipeline state depending on that already existing desired
state.

Original author: suresh-khurdiya-epam <skhurdiya.contractor@libertyglobal.com>

* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::setRate): Pause/unpause the pipeline according to the state of playbackRate.

Canonical link: https://commits.webkit.org/273125@main